Update tests according to new architecture. Also, filter a warning in norm_cross_corr.
